Conflicts:
	src/Intro.cpp
	src/Ship.cpp
	src/Ship.h
	src/gameui/ModelSpinner.cpp
	win32/vc2012/pioneer.vcxproj.filters
	win32/vc2013Express/pioneer.vcxproj.filters

+++ src/gui/GuiLabel.h
@@ -1,4 +1,4 @@
-// Copyright © 2008-2013 Pioneer Developers. See AUTHORS.txt for details
+// Copyright © 2008-2014 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
 #ifndef _GUILABEL_H
@@ -7,6 +7,7 @@
 #include "GuiWidget.h"
 #include "GuiTextLayout.h"
 #include <string>
+#include <SDL_stdinc.h>
 
 namespace Text { class TextureFont; }
 
@@ -21,7 +22,7 @@ namespace Gui {
 		void SetText(const char *text);
 		void SetText(const std::string &text);
 		Label *Shadow(bool isOn) { m_shadow = isOn; return this; }
-		Label *Color(float r, float g, float b);
+		Label *Color(Uint8 r, Uint8 g, Uint8 b);
 		Label *Color(const ::Color &);
 	private:
 		void Init(const std::string &text, TextLayout::ColourMarkupMode colourMarkupMode);
